yopedia: homepage declutter — drop logo glyph, graph, status badge; topics → /wiki
- Nav brand: replace the node-constellation Logo glyph with a plain "yopedia"
  wordmark.
- Homepage: remove the substrate graph (HomeGraph) and make "the trail"
  full-width; remove the "Connected: <provider>" StatusBadge (the same status
  already lives on /settings); move "browse by topic" off the homepage.
- /wiki (browse): add "browse by topic" tag chips (computed from the current
  scope, preserve scope in the link) — this is where browsing belongs.
