+++ src/components/NavHeader.tsx
@@ -6,7 +6,6 @@ import { useState, useEffect, useRef } from "react";
 import { Show, SignInButton, UserButton, useUser } from "@clerk/nextjs";
 import { GlobalSearch } from "./GlobalSearch";
 import { ThemeToggle } from "./ThemeToggle";
-import { Logo } from "./Logo";
 import { isOwnerHandle } from "@/lib/owner";
 
 // Primary actions — shown to everyone, always in the bar.
@@ -83,8 +82,11 @@ export function NavHeader() {
   return (
     <header className="sticky top-0 z-50 bg-background border-b border-border shadow-sm">
       <nav aria-label="Main navigation" className="mx-auto flex h-14 max-w-5xl items-center justify-between px-6">
-        <Link href="/" className="hover:opacity-90 transition-opacity">
-          <Logo />
+        <Link
+          href="/"
+          className="text-lg font-bold tracking-tight text-foreground hover:opacity-90 transition-opacity"
+        >
+          yopedia
         </Link>
 
         {/* Desktop nav */}
